pyk/src/: rename kast.kinner => kast.inner, kast.kouter => kast.outer
