+++ src/integration_tests/kompiled_test.py
@@ -5,8 +5,8 @@ from tempfile import mkdtemp
 from typing import Iterable, Optional
 from unittest import TestCase
 
-from pyk.kast.kinner import KInner
-from pyk.kast.kouter import KDefinition
+from pyk.kast.inner import KInner
+from pyk.kast.outer import KDefinition
 from pyk.ktool import KompileBackend, kompile
 from pyk.prelude.ml import is_top
