+++ src/integration_tests/test_kprint.py
@@ -1,4 +1,4 @@
-from pyk.kast.kinner import KApply, KSequence, KToken, KVariable
+from pyk.kast.inner import KApply, KSequence, KToken, KVariable
 from pyk.kast.manip import remove_attrs
 from pyk.ktool import KompileBackend
 from pyk.ktool.kprint import SymbolTable
