+++ src/integration_tests/test_kprove.py
@@ -1,7 +1,7 @@
 from pyk.cterm import CTerm
-from pyk.kast.kinner import KApply, KAtt, KSequence, KToken, KVariable
-from pyk.kast.kouter import KClaim, KRule
+from pyk.kast.inner import KApply, KAtt, KSequence, KToken, KVariable
 from pyk.kast.manip import get_cell
+from pyk.kast.outer import KClaim, KRule
 from pyk.ktool import KompileBackend
 from pyk.ktool.kprint import SymbolTable
 from pyk.prelude.kbool import BOOL
