+++ src/pyk/coverage.py
@@ -1,8 +1,8 @@
 from os import PathLike
 from typing import Iterable, List
 
-from .kast.kinner import KApply, KRewrite, KSequence
-from .kast.kouter import KDefinition, KRule, read_kast_definition
+from .kast.inner import KApply, KRewrite, KSequence
+from .kast.outer import KDefinition, KRule, read_kast_definition
 
 
 def get_rule_by_id(definition: KDefinition, rule_id: str) -> KRule:
