+++ src/pyk/kcfg.py
@@ -10,9 +10,9 @@ from typing import Any, Container, Dict, Iterable, List, Mapping, Optional, Sequ
 from graphviz import Digraph
 
 from .cterm import CTerm, build_claim, build_rule
-from .kast.kinner import KInner, Subst
-from .kast.kouter import KClaim, KRule
+from .kast.inner import KInner, Subst
 from .kast.manip import ml_pred_to_bool, mlAnd, remove_source_attributes, simplify_bool
+from .kast.outer import KClaim, KRule
 from .ktool import KPrint
 from .utils import add_indent, compare_short_hashes, shorten_hash
