+++ src/tests/mock_kprint.py
@@ -1,4 +1,4 @@
-from pyk.kast.kouter import KDefinition, KFlatModule
+from pyk.kast.outer import KDefinition, KFlatModule
 from pyk.ktool.kprint import KPrint
