+++ src/tests/test_count_vars.py
@@ -1,7 +1,7 @@
 from typing import Final, Mapping, Tuple
 from unittest import TestCase
 
-from pyk.kast.kinner import KInner
+from pyk.kast.inner import KInner
 from pyk.kast.manip import count_vars
 
 from .utils import a, b, c, f, g, h, x, y, z
