+++ src/tests/test_pretty_print_kast.py
@@ -1,8 +1,8 @@
 from typing import Final, Tuple
 from unittest import TestCase
 
-from pyk.kast.kinner import KApply, KAst, KAtt, KLabel, KSort, KVariable
-from pyk.kast.kouter import KNonTerminal, KProduction, KRule, KTerminal
+from pyk.kast.inner import KApply, KAst, KAtt, KLabel, KSort, KVariable
+from pyk.kast.outer import KNonTerminal, KProduction, KRule, KTerminal
 from pyk.ktool.kprint import SymbolTable, pretty_print_kast, unparser_for_production
 from pyk.prelude.kbool import TRUE
