+++ src/tests/utils.py
@@ -1,4 +1,4 @@
-from pyk.kast.kinner import KApply, KLabel, KVariable
+from pyk.kast.inner import KApply, KLabel, KVariable
 
 a, b, c = map(KApply, ('a', 'b', 'c'))
 x, y, z = map(KVariable, ('x', 'y', 'z'))
